Various opcode tweaks.

+++ lib/com/cowlark/luje/FieldBench.java
@@ -9,8 +9,8 @@ package com.cowlark.luje;
 
 public class FieldBench
 {
-	public static long TIME = 2000;
-	public static long STEP = 10000;
+	public static int TIME = 2000;
+	public static int STEP = 10000;
 
 	public static abstract class Benchmark
 	{
